Update convInteger_bias_test
Need to set this to be the same time for both input and bias for the data as this doesn't operate the same as regular Conv

+++ test/onnx/gen_onnx.py
@@ -1533,7 +1533,7 @@ def conv_relu_maxpool_x2_test():
 def convinteger_bias_test():
     x = helper.make_tensor_value_info('0', TensorProto.INT8, [1, 3, 32, 32])
     y = helper.make_tensor_value_info('1', TensorProto.INT8, [1, 3, 5, 5])
-    z = helper.make_tensor_value_info('2', TensorProto.INT32, [1])
+    z = helper.make_tensor_value_info('2', TensorProto.INT8, [1])
     out = helper.make_tensor_value_info('3', TensorProto.INT32, [1, 2, 28, 28])
 
     node = onnx.helper.make_node('ConvInteger',
